version-b-logic-apps/requirements.txt
azure-functions>=1.18.0
azure-data-tables>=12.4.0
